requirements.txt
python-telegram-bot==20.0
aniso8601==9.0.1
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask==2.1.3
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.7.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
pyproject-toml==0.0.10
pyrsistent==0.18.1
pytz==2022.1
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.39
toml==0.10.2
urllib3==1.26.10
Werkzeug==2.2.0

